# api/openapi.json
{
  "components": {
    "schemas": {
      "ApiError": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ApiErrorCode"
          },
          "fields": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "propertyNames": {
              "type": "string"
            },
            "type": [
              "object",
              "null"
            ]
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
      "ApiErrorCode": {
        "description": "Closed, stable machine-readable failure codes shared by every public HTTP\nresponse. Endpoint response declarations constrain which subset is\napplicable, while the runtime wire type cannot emit an unregistered code.",
        "enum": [
          "request_forbidden",
          "unsupported_media_type",
          "payload_too_large",
          "malformed_json",
          "invalid_request",
          "resource_not_found",
          "method_not_allowed",
          "revision_conflict",
          "command_not_found",
          "command_state_conflict",
          "profile_switch_conflict",
          "serial_connection_failed",
          "camera_unavailable",
          "screenshot_unavailable",
          "destination_conflict",
          "notification_unsupported",
          "dynamic_config_unavailable",
          "unsupported_platform",
          "persistence_failed",
          "backend_unavailable",
          "internal_error"
        ],
        "type": "string"
      },
...
      "CameraSelector": {
        "oneOf": [
          {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          {
            "type": "string"
          }
        ]
      },
...
      "DecimalString": {
        "description": "An exact non-negative integer wire value that never depends on JavaScript\n`Number` precision.",
        "example": "42",
        "pattern": "^(0|[1-9][0-9]*)$",
        "type": "string"
      },
...
      "EmptyRequest": {
        "additionalProperties": false,
        "type": "object"
      },
      "ErrorEnvelope": {
        "additionalProperties": false,
        "description": "Common secret-safe error envelope.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiError"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "ImageFormat": {
        "enum": [
          "png",
          "jpeg"
        ],
        "type": "string"
      },
...
      "NormalizedRegion": {
        "additionalProperties": false,
        "properties": {
          "height": {
            "format": "double",
            "maximum": 1,
            "minimum": 0,
            "type": "number"
          },
          "width": {
            "format": "double",
            "maximum": 1,
            "minimum": 0,
            "type": "number"
          },
          "x": {
            "format": "double",
            "maximum": 1,
            "minimum": 0,
            "type": "number"
          },
          "y": {
            "format": "double",
            "maximum": 1,
            "minimum": 0,
            "type": "number"
          }
        },
        "required": [
          "x",
          "y",
          "width",
          "height"
        ],
        "type": "object"
      },
...
      "ScreenshotCaptures": {
        "additionalProperties": false,
        "properties": {
          "filename": {
            "type": [
              "string",
              "null"
            ]
          },
          "format": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ImageFormat"
              }
            ]
          },
          "overwrite": {
            "type": "boolean"
          },
          "region": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/NormalizedRegion"
              }
            ]
          }
        },
        "type": "object"
      },
      "ScreenshotDownload": {
        "additionalProperties": false,
        "properties": {
          "filename": {
            "type": [
              "string",
              "null"
            ]
          },
          "format": {
            "$ref": "#/components/schemas/ImageFormat"
          },
          "region": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/NormalizedRegion"
              }
            ]
          }
        },
        "required": [
          "format"
        ],
        "type": "object"
      },
      "ScreenshotPath": {
        "additionalProperties": false,
        "properties": {
          "format": {
            "$ref": "#/components/schemas/ImageFormat"
          },
          "overwrite": {
            "type": "boolean"
          },
          "path": {
            "type": "string"
          },
          "region": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/NormalizedRegion"
              }
            ]
          }
        },
        "required": [
          "path",
          "format"
        ],
        "type": "object"
      },
      "ScreenshotRequest": {
        "discriminator": {
          "propertyName": "destination"
        },
        "oneOf": [
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/ScreenshotCaptures"
              },
              {
                "additionalProperties": false,
                "properties": {
                  "destination": {
                    "enum": [
                      "captures"
                    ],
                    "type": "string"
                  }
                },
                "required": [
                  "destination"
                ],
                "type": "object"
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/ScreenshotPath"
              },
              {
                "additionalProperties": false,
                "properties": {
                  "destination": {
                    "enum": [
                      "path"
                    ],
                    "type": "string"
                  }
                },
                "required": [
                  "destination"
                ],
                "type": "object"
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/ScreenshotDownload"
              },
              {
                "additionalProperties": false,
                "properties": {
                  "destination": {
                    "enum": [
                      "download"
                    ],
                    "type": "string"
                  }
                },
                "required": [
                  "destination"
                ],
                "type": "object"
              }
            ]
          }
        ]
      },
      "SerialControlRequest": {
        "discriminator": {
          "propertyName": "action"
        },
        "oneOf": [
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/EmptyRequest"
              },
              {
                "additionalProperties": false,
                "properties": {
                  "action": {
                    "enum": [
                      "connect"
                    ],
                    "type": "string"
                  }
                },
                "required": [
                  "action"
                ],
                "type": "object"
              }
            ]
          },
          {
            "allOf": [
              {
                "$ref": "#/components/schemas/EmptyRequest"
              },
              {
                "additionalProperties": false,
                "properties": {
                  "action": {
                    "enum": [
                      "disconnect"
                    ],
                    "type": "string"
                  }
                },
                "required": [
                  "action"
                ],
                "type": "object"
              }
            ]
          }
        ]
      },
...
      "Success_OperationResult": {
        "additionalProperties": false,
        "description": "Common successful JSON envelope.",
        "properties": {
          "data": {
            "additionalProperties": false,
            "properties": {
              "changed": {
                "type": "boolean"
              },
              "revision": {
                "$ref": "#/components/schemas/DecimalString"
              }
            },
            "required": [
              "changed",
              "revision"
            ],
            "type": "object"
          }
        },
        "required": [
          "data"
        ],
        "type": "object"
      },
      "Success_SavedScreenshot": {
        "additionalProperties": false,
        "description": "Common successful JSON envelope.",
        "properties": {
          "data": {
            "additionalProperties": false,
            "properties": {
              "display_path": {
                "type": "string"
              },
              "format": {
                "$ref": "#/components/schemas/ImageFormat"
              }
            },
            "required": [
              "display_path",
              "format"
            ],
            "type": "object"
          }
        },
        "required": [
          "data"
        ],
        "type": "object"
      },
...
      "Success_Vec_CameraDevice": {
        "additionalProperties": false,
        "description": "Common successful JSON envelope.",
        "properties": {
          "data": {
            "items": {
              "additionalProperties": false,
              "properties": {
                "available": {
                  "type": "boolean"
                },
                "label": {
                  "type": "string"
                },
                "selector": {
                  "$ref": "#/components/schemas/CameraSelector"
                }
              },
              "required": [
                "selector",
                "label",
                "available"
              ],
              "type": "object"
            },
            "type": "array"
          }
        },
        "required": [
          "data"
        ],
        "type": "object"
      },
      "Success_Vec_SerialPort": {
        "additionalProperties": false,
        "description": "Common successful JSON envelope.",
        "properties": {
          "data": {
            "items": {
              "additionalProperties": false,
              "properties": {
                "available": {
                  "type": "boolean"
                },
                "label": {
                  "type": "string"
                },
                "selector": {
                  "type": "string"
                }
              },
              "required": [
                "selector",
                "label",
                "available"
              ],
              "type": "object"
            },
            "type": "array"
          }
        },
        "required": [
          "data"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Closed REST and WebSocket contracts for PokeCon",
    "license": {
      "identifier": "MIT",
      "name": "MIT"
    },
    "title": "PokeCon API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/camera/retry": {
      "post": {
        "operationId": "retry_camera",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmptyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success_OperationResult"
                }
              }
            },
            "description": "Camera retry applied or a no-op"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Camera retry failure"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Camera backend failure"
          }
        },
        "tags": [
          "devices"
        ]
      }
    },
    "/api/camera/screenshot": {
      "post": {
        "operationId": "screenshot",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScreenshotRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success_SavedScreenshot"
                }
              },
              "image/jpeg": {
                "schema": {
                  "items": {
                    "format": "int32",
                    "minimum": 0,
                    "type": "integer"
                  },
                  "type": "array"
                }
              },
              "image/png": {
                "schema": {
                  "items": {
                    "format": "int32",
                    "minimum": 0,
                    "type": "integer"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Saved screenshot metadata or downloaded image bytes",
            "headers": {
              "Content-Disposition": {
                "description": "Present for attachment responses",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No frame or incompatible UI mode"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Invalid crop, name, path, or format"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Screenshot failure"
          }
        },
        "tags": [
          "devices"
        ]
      }
    },
...
    "/api/devices/cameras": {
      "get": {
        "operationId": "get_cameras",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success_Vec_CameraDevice"
                }
              }
            },
            "description": "Fresh camera enumeration"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Camera enumeration failure"
          }
        },
        "tags": [
          "devices"
        ]
      }
    },
    "/api/devices/serial-ports": {
      "get": {
        "operationId": "get_serial_ports",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success_Vec_SerialPort"
                }
              }
            },
            "description": "Fresh serial-port enumeration"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Serial enumeration failure"
          }
        },
        "tags": [
          "devices"
        ]
      }
    },
...
    "/api/serial/control": {
      "post": {
        "operationId": "control_serial",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SerialControlRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success_OperationResult"
                }
              }
            },
            "description": "Serial transition applied or a no-op"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Serial connection failure"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Invalid serial request"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Serial backend failure"
          }
        },
        "tags": [
          "devices"
        ]
      }
    },
...
  },
...
}